Add jsdocs for custom exceptions.
