+++ fileHub/js-app/app/services/request-error.js
@@ -1,5 +1,6 @@
 /**
- *
+ * Error with status codes that indicate that there was likely an error in the request
+ * which prevented the server from being able to process it.
  */
 export class RequestError extends Error {
   /**
