+++ fileHub/js-app/app/services/server-error.js
@@ -1,5 +1,5 @@
 /**
- *
+ * Error that throws when response status equals 500.
  */
 export class ServerError extends Error {
   /**
